requirements.txt
network
machine
micropython-cpython-ure